Refactor views folder structure
