+++ src/views/Profile/AditionalForm.tsx
@@ -1,10 +1,10 @@
 import { Box } from '@impact-market/ui';
-import { selectCurrentUser } from '../../../state/slices/auth';
+import { selectCurrentUser } from '../../state/slices/auth';
 import { useForm, useFormState } from "react-hook-form";
-import { usePrismicData } from '../../../libs/Prismic/components/PrismicDataProvider';
+import { usePrismicData } from '../../libs/Prismic/components/PrismicDataProvider';
 import { useSelector } from 'react-redux';
 import FormActions from './FormActions';
-import Input from '../../../components/Input';
+import Input from '../../components/Input';
 import React, { useEffect } from "react";
 
 const Form = ({ onSubmit }: any) => {
