+++ src/views/Profile/ContactForm.tsx
@@ -1,13 +1,13 @@
 import { Box } from '@impact-market/ui';
-import { selectCurrentUser } from '../../../state/slices/auth';
+import { selectCurrentUser } from '../../state/slices/auth';
 import { useForm, useFormState } from "react-hook-form";
-import { usePrismicData } from '../../../libs/Prismic/components/PrismicDataProvider';
+import { usePrismicData } from '../../libs/Prismic/components/PrismicDataProvider';
 import { useSelector } from 'react-redux';
 import FormActions from './FormActions';
-import Input from '../../../components/Input';
+import Input from '../../components/Input';
 import React, { useEffect } from "react";
-import RichText from '../../../libs/Prismic/components/RichText';
-import useTranslations from '../../../libs/Prismic/hooks/useTranslations';
+import RichText from '../../libs/Prismic/components/RichText';
+import useTranslations from '../../libs/Prismic/hooks/useTranslations';
 
 const Form = ({ onSubmit }: any) => {
     const auth = useSelector(selectCurrentUser);
