+++ src/views/Profile/FormActions.tsx
@@ -1,6 +1,6 @@
 import { Box, Button, Col, Divider, Row } from '@impact-market/ui';
 import React from "react";
-import String from '../../../libs/Prismic/components/String';
+import String from '../../libs/Prismic/components/String';
 
 const FormActions = ({ handleCancel, isSubmitting }: any) => {    
     return (
