+++ src/views/Settings/index.tsx
@@ -4,7 +4,7 @@ import { setUser } from '../../state/slices/auth';
 import { useDispatch } from 'react-redux';
 import { usePrismicData } from '../../libs/Prismic/components/PrismicDataProvider';
 import { useUpdateUserMutation } from '../../api/user';
-import BasicForm from './components/BasicForm';
+import BasicForm from './BasicForm';
 import React from 'react';
 import RichText from '../../libs/Prismic/components/RichText';
